+++ jsoar-core/src/main/java/org/jsoar/kernel/SoarProperties.java
@@ -113,77 +113,77 @@ public class SoarProperties
      * 
      * @see DecisionCycle
      */
-    public static final PropertyKey<Integer> D_CYCLE_COUNT = counter("d_cycle_count");
+    public static final PropertyKey<Long> D_CYCLE_COUNT = counter("d_cycle_count");
     
     /**
      * <p>agent.h::e_cycle_count
      * 
      * @see DecisionCycle
      */
-    public static final PropertyKey<Integer> E_CYCLE_COUNT = counter("e_cycle_count");
+    public static final PropertyKey<Long> E_CYCLE_COUNT = counter("e_cycle_count");
 
     /**
      * <p>agent.h::decision_phases_count
      * 
      * @see DecisionCycle
      */
-    public static final PropertyKey<Integer> DECISION_PHASES_COUNT = counter("decision_phases_count");
+    public static final PropertyKey<Long> DECISION_PHASES_COUNT = counter("decision_phases_count");
 
     /**
      * <p>agent.h::inner_e_cycle_count
      * 
      * @see DecisionCycle
      */
-    public static final PropertyKey<Integer> INNER_E_CYCLE_COUNT = counter("inner_e_cycle_count");
+    public static final PropertyKey<Long> INNER_E_CYCLE_COUNT = counter("inner_e_cycle_count");
 
     /**
      * <p>agent.h::pe_cycle_count
      * 
      * @see DecisionCycle
      */
-    public static final PropertyKey<Integer> PE_CYCLE_COUNT = counter("pe_cycle_count");
+    public static final PropertyKey<Long> PE_CYCLE_COUNT = counter("pe_cycle_count");
 
     /**
      * <p>agent.h::gent.h:367:production_firing_count
      * 
      * @see RecognitionMemory
      */
-    public static final PropertyKey<Integer> PRODUCTION_FIRING_COUNT = counter("production_firing_count");
+    public static final PropertyKey<Long> PRODUCTION_FIRING_COUNT = counter("production_firing_count");
 
     /**
      * <p>agent.h::wme_addition_count
      * 
      * @see WorkingMemory
      */
-    public static final PropertyKey<Integer> WME_ADDITION_COUNT = counter("wme_addition_count");
+    public static final PropertyKey<Long> WME_ADDITION_COUNT = counter("wme_addition_count");
     
     /**
      * <p>agent.h::wme_removal_count
      * 
      * @see WorkingMemory
      */
-    public static final PropertyKey<Integer> WME_REMOVAL_COUNT = counter("wme_removal_count");
+    public static final PropertyKey<Long> WME_REMOVAL_COUNT = counter("wme_removal_count");
 
     /**
      * <p>agent.h::max_wm_size
      * 
      * @see WorkingMemory
      */
-    public static final PropertyKey<Integer> MAX_WM_SIZE = counter("max_wm_size");
+    public static final PropertyKey<Long> MAX_WM_SIZE = counter("max_wm_size");
     
     /**
      * <p>agent.h::cumulative_wm_size
      * 
      * @see WorkingMemory
      */
-    public static final PropertyKey<Integer> CUMULATIVE_WM_SIZE = counter("cumulative_wm_size");
+    public static final PropertyKey<Long> CUMULATIVE_WM_SIZE = counter("cumulative_wm_size");
     
     /**
      * <p>agent.h::num_wm_sizes_accumulated
      * 
      * @see WorkingMemory
      */
-    public static final PropertyKey<Integer> NUM_WM_SIZES_ACCUMULATED = counter("num_wm_sizes_accumulated");
+    public static final PropertyKey<Long> NUM_WM_SIZES_ACCUMULATED = counter("num_wm_sizes_accumulated");
     
     /**
      * True if the agent is currently running. This property is not set on a raw 
@@ -199,8 +199,8 @@ public class SoarProperties
      */
     public static final PropertyKey<WaitInfo> WAIT_INFO = PropertyKey.builder("wait_info", WaitInfo.class).readonly(true).defaultValue(WaitInfo.NOT_WAITING).build();
     
-    private static PropertyKey<Integer> counter(String name)
+    private static PropertyKey<Long> counter(String name)
     {
-        return PropertyKey.builder(name, Integer.class).boundable(false).readonly(true).defaultValue(0).build();
+        return PropertyKey.builder(name, Long.class).boundable(false).readonly(true).defaultValue(0L).build();
     }
 }
